requirements.txt
# --- AI & Data Core Requirements ---
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.7.4.5
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==24.2
pandas==2.3.2
pillow==10.4.0
protobuf==5.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1

# --- Full Stack Requirements ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.36.0
python-multipart==0.0.9